Rework View class structure to better reflect behaviours.
This moves core view functions into an abstract BaseView class.

- Twig and PHP views extend this.
- 'View' is deprecated, but extends PhpView for backwards-compat.
